requirements.txt
img2pdf
pillow
fpdf
fpdf2
PyPDF2
tkinter